CORE-518 Parse numeric strings with BigDecimal
git-svn-id: http://liquibase.jira.com/svn/CORE/trunk@1268 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ liquibase-core/src/main/java/liquibase/database/typeconversion/core/AbstractTypeConverter.java
@@ -10,6 +10,7 @@ import liquibase.exception.DateParseException;
 import liquibase.exception.UnexpectedLiquibaseException;
 import liquibase.change.ColumnConfig;
 
+import java.math.BigDecimal;
 import java.text.ParseException;
 import java.sql.Types;
 import java.math.BigInteger;
@@ -104,9 +105,9 @@ public abstract class AbstractTypeConverter implements TypeConverter {
                 if (decimalDigits == 0) {
                     return new Integer(value);
                 }
-                return new Double(value);
+                return new BigDecimal(value);
             } else if (dataType == Types.DOUBLE || dataType == Types.NUMERIC) {
-                return new Double(value);
+                return new BigDecimal(value);
             } else if (dataType == Types.FLOAT) {
                 return new Float(value);
             } else if (dataType == Types.INTEGER) {
